requirements.txt
django~=5.0.6
django-rest-framework
whitenoise
psycopg2
dj-database-url
django-heroku
gunicorn
python-decouple
numpy
django-cors-headers
requests
Pillow

djangorestframework~=3.15.1
pandas~=2.2.0